Move ImportPerformanceLogger class to separate file
